@@ -49,6 +49,7 @@
     "@types/fs-extra": "8.0.0",
     "@types/globby": "9.1.0",
     "@types/ini": "1.3.30",
+    "@types/internal-ip": "4.1.0",
     "@types/mime": "2.0.1",
     "@types/mocha": "5.2.7",
     "@types/node": "12.6.9",
@@ -65,8 +66,9 @@
     "should": "13.2.3",
     "sinon": "7.5.0",
     "ts-node": "8.3.0",
-    "tslint": "5.18.0",
-    "typescript": "3.5.3"
+    "tslint": "6.0.0",
+    "typescript": "3.8.3",
+    "webpack": "4.42.0"
   },
   "dependencies": {
     "@signageos/sdk": "1.3.1-beta.1",
@@ -83,14 +85,13 @@
     "fs-extra": "8.1.0",
     "globby": "10.0.1",
     "ini": "1.3.5",
+    "internal-ip": "5.0.0",
     "mime": "2.4.4",
     "node-fetch": "2.6.0",
     "parse-gitignore": "1.0.1",
     "prompts": "2.1.0",
     "read-chunk": "3.2.0",
     "semver": "7.1.3",
-    "serve-static": "1.14.1",
-    "webpack": "4.42.0",
-    "webpack-dev-server": "3.8.0"
+    "serve-static": "1.14.1"
   }
 }
